back/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
colorama==0.4.6
easyocr==1.6.2
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
imageio==2.35.1
lazy_loader==0.4
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
networkx==3.1
ninja==1.13.0
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
opencv-python-headless==4.5.4.60
packaging==25.0
Pillow==9.5.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.11
python-bidi==0.6.6
python-multipart==0.0.20
PyWavelets==1.4.1
PyYAML==6.0.2
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.7
sniffio==1.3.1
starlette==0.44.0
tifffile==2023.7.10
torch==1.7.1
torchvision==0.8.2
typing_extensions==4.13.2
uvicorn==0.33.0
